refactor: articleList
