+++ ssafit-backend/src/main/resources/application.properties
@@ -1,9 +1,9 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/ssafit_db?serverTimezone=UTC
-spring.datasource.username=jason9865
-spring.datasource.password=Tmdgh9275$&
-#spring.datasource.username=ssafy
-#spring.datasource.password=ssafy
+#spring.datasource.username=jason9865
+#spring.datasource.password=Tmdgh9275$&
+spring.datasource.username=ssafy
+spring.datasource.password=ssafy
 
 mybatis.mapper-locations=classpath:mappers/*.xml
 mybatis.type-aliases-package=com.ssafit.*.model.dto, com.ssafit.*.model.entity
